Content Model step 0
